Fix phpunit
