+++ composer.json
@@ -87,7 +87,7 @@
         "behat/mink-selenium2-driver": "^1.3.1",
         "phing/phing": "^2.16",
         "symfony/dotenv": "^4.0",
-        "symfony/phpunit-bridge": "^4.0",
+        "symfony/phpunit-bridge": "^5.1",
         "symfony/profiler-pack": "^1.0"
     },
