+++ tests/Sonata/Bundle/QABundle/SonataAdminExplainCommandTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Sonata\Bundle\QABundle\Tests;
+namespace Tests\Sonata\Bundle\QABundle;
 
 class SonataAdminExplainCommandTest extends CommandTestCase
 {
