+++ tests/Sonata/Bundle/QABundle/SonataNotificationListQueuesCommandTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Sonata\Bundle\QABundle\Tests;
+namespace Tests\Sonata\Bundle\QABundle;
 
 class SonataNotificationListQueuesCommandTest extends CommandTestCase
 {
@@ -21,6 +21,6 @@ class SonataNotificationListQueuesCommandTest extends CommandTestCase
 
         $output = $this->runCommand($client, 'sonata:notification:list-queues');
 
-        $this->assertContains('The backend class Sonata\\NotificationBundle\\Backend\\PostponeRuntimeBackend does not provide multiple queues.', $output);
+        $this->assertStringContainsString('The backend class Sonata\\NotificationBundle\\Backend\\PostponeRuntimeBackend does not provide multiple queues.', $output);
     }
 }
