+++ tests/Sonata/Bundle/QABundle/SonataUserTwoStepVerificationCommandTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Sonata\Bundle\QABundle\Tests;
+namespace Tests\Sonata\Bundle\QABundle;
 
 class SonataUserTwoStepVerificationCommandTest extends CommandTestCase
 {
@@ -30,13 +30,13 @@ class SonataUserTwoStepVerificationCommandTest extends CommandTestCase
 
         $output = $this->runCommand($client, 'sonata:user:two-step-verification --reset secure');
 
-        $this->assertContains('Url : https://chart.googleapis.com/', $output);
+        $this->assertStringContainsString('Url : https://chart.googleapis.com/', $output);
 
         $user = $client->getContainer()->get('fos_user.user_manager')->findUserBy([
             'username' => 'secure',
         ]);
 
-        $this->assertContains($user->getTwoStepVerificationCode(), $output);
+        $this->assertStringContainsString($user->getTwoStepVerificationCode(), $output);
     }
 
     public function testGenerateOnGenerateUser()
@@ -49,6 +49,6 @@ class SonataUserTwoStepVerificationCommandTest extends CommandTestCase
             'username' => 'secure',
         ])->getTwoStepVerificationCode();
 
-        $this->assertContains($code, $output);
+        $this->assertStringContainsString($code, $output);
     }
 }
